backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
anthropic==0.28.0
httpx==0.27.0
beautifulsoup4==4.12.3
PyMuPDF==1.24.3
pandas==2.2.2
python-multipart==0.0.9
pydantic==2.7.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
aiofiles==23.2.1
